+++ plugins/catalog-backend/src/next/database/types.ts
@@ -17,11 +17,17 @@
 import { Entity, EntityRelationSpec } from '@backstage/catalog-model';
 import { JsonObject } from '@backstage/config';
 import { Transaction } from '../../database/types';
+import { DeferredEntity } from '../processing/types';
 
-export type AddUnprocessedEntitiesOptions = {
-  entityRef: string;
-  entities: Entity[];
-};
+export type AddUnprocessedEntitiesOptions =
+  | {
+      sourceEntityRef: string;
+      entities: DeferredEntity[];
+    }
+  | {
+      sourceKey: string;
+      entities: DeferredEntity[];
+    };
 
 export type AddUnprocessedEntitiesResult = {};
 
@@ -31,7 +37,8 @@ export type UpdateProcessedEntityOptions = {
   state?: Map<string, JsonObject>;
   errors?: string;
   relations: EntityRelationSpec[];
-  deferredEntities: Entity[];
+  deferredEntities: DeferredEntity[];
+  locationKey?: string;
 };
 
 export type UpdateProcessedEntityErrorsOptions = {
@@ -48,6 +55,7 @@ export type RefreshStateItem = {
   lastDiscoveryAt: string; // remove?
   state: Map<string, JsonObject>;
   errors?: string;
+  locationKey?: string;
 };
 
 export type GetProcessableEntitiesResult = {
@@ -57,23 +65,18 @@ export type GetProcessableEntitiesResult = {
 export type ReplaceUnprocessedEntitiesOptions =
   | {
       sourceKey: string;
-      items: Entity[];
+      items: DeferredEntity[];
       type: 'full';
     }
   | {
       sourceKey: string;
-      added: Entity[];
-      removed: Entity[];
+      added: DeferredEntity[];
+      removed: DeferredEntity[];
       type: 'delta';
     };
 
 export interface ProcessingDatabase {
   transaction<T>(fn: (tx: Transaction) => Promise<T>): Promise<T>;
-
-  addUnprocessedEntities(
-    tx: Transaction,
-    options: AddUnprocessedEntitiesOptions,
-  ): Promise<void>;
 
   replaceUnprocessedEntities(
     txOpaque: Transaction,