+++ packages/catalog-client/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @backstage/catalog-client
 
+## 0.3.15
+
+### Patch Changes
+
+- ca080cab8: Don't crash if the entities response doesn't include the entities name and kind
+
 ## 0.3.14
 
 ### Patch Changes